Add files via upload
This playback script assumed that you have an openlog in normal mode (IE not playback mode which is set by changing the config file on the SD card) connected to an UNO, which is also connected to the eggbot serial pins (TTL RX/TX)

This sketch was written in record time. As such, it has some flaws. 

1.) File must be named 1.ebb
2.) File must start with a newline
3.) Newlines must be mac style -- '\r' only 
4.) File must NOT end with a blank line
5.) All button query commands must be removed as the arduino doesn't answer that prompt correctly (just issues OK)
6.) As written it repeats the file forever as this was made for a window display. Should be trivial to have it just not repeat the file at the end. 
7.) Use included ebb file for an example (rename to 0.ebb)